+++ CHANGELOG.md
@@ -4,9 +4,18 @@ All notable changes in keycloak-subject-id-attribute-mapper will be documented i
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-It is based on Keycloak version 22.
+
+## [2.0.0] - 2024-10-23
+
+### Changed
+- It is based on Keycloak version 26.5.x
+- Change to support multiple Identity Providers mappers
+
+### Added
+- Support for User attribute Template Importer
 
 ## [1.1.0] - 2024-10-23
+It is based on Keycloak version 22.
 
 ### Added
 - Check for unique subject id